+++ ports/zephyr/mpconfigport.h
@@ -139,6 +139,3 @@ typedef long mp_off_t;
 // extra built in names to add to the global namespace
 #define MICROPY_PORT_BUILTINS \
     { MP_ROM_QSTR(MP_QSTR_open), MP_ROM_PTR(&mp_builtin_open_obj) },
-
-#define MICROPY_BEGIN_ATOMIC_SECTION irq_lock
-#define MICROPY_END_ATOMIC_SECTION irq_unlock
